Splict strcuts into types.go to request.go and response.go
